cca: move RSI related definitions to RSI crate
