+++ openhcl/hcl/Cargo.toml
@@ -18,6 +18,7 @@ x86defs.workspace = true
 inspect.workspace = true
 user_driver.workspace = true
 aarch64defs.workspace = true
+rsi.workspace = true
 
 anyhow.workspace = true
 arrayvec.workspace = true
