+++ openhcl/hcl/src/ioctl/cca.rs
@@ -15,17 +15,17 @@ use crate::ioctl::ioctls::mshv_realm_config;
 use crate::ioctl::ioctls::mshv_rsi_set_mem_perm;
 use crate::ioctl::ioctls::mshv_rsi_sysreg_write;
 use crate::ioctl::ioctls::{hcl_realm_config, hcl_rsi_set_mem_perm, hcl_rsi_sysreg_write};
-use crate::protocol::RSI_PLANE_ENTER_FLAGS_TRAP_SIMD;
-use crate::protocol::RSI_PLANE_GIC_NUM_LRS;
-use crate::protocol::RSI_PLANE_NR_GPRS;
-use crate::protocol::cca_rsi_plane_entry;
-use crate::protocol::cca_rsi_plane_exit;
-use crate::protocol::cca_rsi_plane_run;
 use aarch64defs::SystemReg;
 use hvdef::HV_PAGE_SIZE;
 use hvdef::HvArm64RegisterName;
 use hvdef::HvRegisterName;
 use hvdef::HvRegisterValue;
+use rsi::RSI_PLANE_ENTER_FLAGS_TRAP_SIMD;
+use rsi::RSI_PLANE_GIC_NUM_LRS;
+use rsi::RSI_PLANE_NR_GPRS;
+use rsi::cca_rsi_plane_entry;
+use rsi::cca_rsi_plane_exit;
+use rsi::cca_rsi_plane_run;
 use sidecar_client::SidecarVp;
 use user_driver::memory::MemoryBlock;
