+++ openhcl/virt_mshv_vtl/src/processor/cca/mod.rs
@@ -17,13 +17,13 @@ use crate::processor::InterceptMessageState;
 use crate::{BackingShared, UhCvmPartitionState, UhCvmVpState, UhPartitionNewParams};
 use aarch64defs::EsrEl2;
 use aarch64defs::SystemReg;
-use hcl::protocol::cca_rsi_plane_exit;
 use hcl::{GuestVtl, ioctl::cca::Cca, ioctl::register};
 use hv1_emulator::hv::ProcessorVtlHv;
 use hv1_emulator::synic::ProcessorSynic;
 use hv1_structs::VtlArray;
 use hvdef::HvRegisterCrInterceptControl;
 use inspect::{Inspect, InspectMut};
+use rsi::cca_rsi_plane_exit;
 use virt::VpIndex;
 use virt::aarch64::vp;
 use virt::io::CpuIo;
