+++ vm/aarch64/rsi/Cargo.toml
@@ -16,6 +16,7 @@ memory_range.workspace = true
 aarch64defs.workspace = true
 bitfield-struct.workspace = true
 open_enum.workspace = true
+zerocopy.workspace = true
 
 tracing = { workspace = true, optional = true }
